refactor(modularization): Phase 5 step 5-S2b — agent_runtime DB carve (task #26)
Move ``AgentTurn`` / ``AgentTimelineEvent`` / ``AgentArtifact`` (and
their three enums — ``AgentTurnStatus`` / ``AgentEventActor`` /
``AgentArtifactType``) into ``aperag/domains/agent_runtime/db``, and
extend the aggregate re-export shim in ``aperag.db.models`` with the
six new symbols alongside the conversation entries from 5-S2a.

Tables are FK-linked by opaque string columns (``chat_id`` /
``turn_id`` / ``bot_id``), so there is no Python-level cross-domain
import — the turn service joins them at query time. ``_random_id``
and ``_enum_column`` helper twins mirror the Phase 3 / 5-S2a pattern;
Phase 6 cleanup consolidates onto ``aperag.db.base``.

``aperag.db.models`` now re-exports 30 symbols post-5-S2b (Phase 3 15
+ conversation 9 + agent_runtime 6). G11 final locks in at 5-S8 after
the evaluation carve in 5-S2c.

Verify:
* ``import aperag.db.models`` + 6 symbol identity check OK
* ``AgentTurn is aperag.domains.agent_runtime.db.models.AgentTurn``
* ``pytest tests/unit_test/`` 671 / 29 skipped
* ``uvx ruff check`` + ``uvx ruff format --check`` clean
* ``Base.metadata.tables`` 37 tables still registered
