+++ aperag/db/models.py
@@ -79,28 +79,6 @@ class Role(str, Enum):
     RO = "ro"
 
 
-class AgentTurnStatus(str, Enum):
-    QUEUED = "QUEUED"
-    RUNNING = "RUNNING"
-    COMPLETED = "COMPLETED"
-    FAILED = "FAILED"
-    CANCELLED = "CANCELLED"
-
-
-class AgentEventActor(str, Enum):
-    AGENT = "agent"
-    TOOL = "tool"
-    SYSTEM = "system"
-
-
-class AgentArtifactType(str, Enum):
-    ANSWER = "answer"
-    REFERENCE_BUNDLE = "reference_bundle"
-    TOOL_RESULT_SUMMARY = "tool_result_summary"
-    SEARCH_RESULT_SUMMARY = "search_result_summary"
-    ERROR_SUMMARY = "error_summary"
-
-
 class ModelServiceProviderStatus(str, Enum):
     ACTIVE = "ACTIVE"
     INACTIVE = "INACTIVE"
@@ -272,67 +250,6 @@ class UserQuota(Base):
     def can_consume(self, amount: int = 1) -> bool:
         """Check if the specified amount can be consumed without exceeding quota"""
         return not self.is_quota_exceeded(amount)
-
-
-class AgentTurn(Base):
-    __tablename__ = "agent_turn"
-    __table_args__ = (
-        UniqueConstraint("chat_id", "client_idempotency_key", name="uq_agent_turn_chat_idempotency"),
-        Index("idx_agent_turn_chat_created", "chat_id", "gmt_created"),
-        Index("idx_agent_turn_user_status", "user", "status"),
-    )
-
-    id = Column(String(24), primary_key=True, default=lambda: "turn" + random_id())
-    chat_id = Column(String(24), nullable=False, index=True)
-    user = Column(String(256), nullable=False, index=True)
-    bot_id = Column(String(24), nullable=False, index=True)
-    request_id = Column(String(64), nullable=False, unique=True, index=True)
-    client_idempotency_key = Column(String(128), nullable=False)
-    status = Column(EnumColumn(AgentTurnStatus), nullable=False, default=AgentTurnStatus.QUEUED, index=True)
-    input_text = Column(Text, nullable=False)
-    model_profile = Column(JSON, default=lambda: {}, nullable=False)
-    error_code = Column(String(128), nullable=True)
-    error_message = Column(Text, nullable=True)
-    answer_artifact_id = Column(String(24), nullable=True, index=True)
-    reference_bundle_artifact_id = Column(String(24), nullable=True, index=True)
-    timeline_cursor = Column(Integer, default=0, nullable=False)
-    gmt_created = Column(DateTime(timezone=True), default=utc_now, nullable=False)
-    gmt_started = Column(DateTime(timezone=True), nullable=True)
-    gmt_finished = Column(DateTime(timezone=True), nullable=True)
-    gmt_updated = Column(DateTime(timezone=True), default=utc_now, nullable=False)
-
-
-class AgentTimelineEvent(Base):
-    __tablename__ = "agent_timeline_event"
-    __table_args__ = (
-        UniqueConstraint("turn_id", "sequence", name="uq_agent_timeline_event_turn_sequence"),
-        Index("idx_agent_timeline_event_turn_timestamp", "turn_id", "timestamp"),
-    )
-
-    id = Column(String(24), primary_key=True, default=lambda: "evt" + random_id())
-    turn_id = Column(String(24), nullable=False, index=True)
-    sequence = Column(Integer, nullable=False)
-    timestamp = Column(DateTime(timezone=True), default=utc_now, nullable=False, index=True)
-    type = Column(String(128), nullable=False, index=True)
-    label = Column(String(128), nullable=True)
-    status = Column(String(64), nullable=True)
-    actor = Column(EnumColumn(AgentEventActor), nullable=False, default=AgentEventActor.SYSTEM)
-    data = Column(JSON, default=lambda: {}, nullable=False)
-    gmt_created = Column(DateTime(timezone=True), default=utc_now, nullable=False)
-
-
-class AgentArtifact(Base):
-    __tablename__ = "agent_artifact"
-    __table_args__ = (Index("idx_agent_artifact_turn_type", "turn_id", "artifact_type"),)
-
-    id = Column(String(24), primary_key=True, default=lambda: "art" + random_id())
-    turn_id = Column(String(24), nullable=False, index=True)
-    artifact_type = Column(EnumColumn(AgentArtifactType), nullable=False, index=True)
-    summary = Column(Text, nullable=True)
-    payload = Column(JSON, default=lambda: {}, nullable=False)
-    storage_ref = Column(Text, nullable=True)
-    gmt_created = Column(DateTime(timezone=True), default=utc_now, nullable=False)
-    gmt_updated = Column(DateTime(timezone=True), default=utc_now, nullable=False)
 
 
 class ApiKey(Base):
@@ -915,6 +832,14 @@ class EvaluationRunItemAttempt(Base):
 # import site has migrated to the canonical per-domain path.
 # ---------------------------------------------------------------------------
 
+from aperag.domains.agent_runtime.db.models import (  # noqa: E402, F401  re-export for back-compat
+    AgentArtifact,
+    AgentArtifactType,
+    AgentEventActor,
+    AgentTimelineEvent,
+    AgentTurn,
+    AgentTurnStatus,
+)
 from aperag.domains.conversation.db.models import (  # noqa: E402, F401  re-export for back-compat
     Bot,
     BotStatus,
